+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/card/CardRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardRepository extends JpaRepository<Card, Integer> {
+public interface CardRepository extends JpaRepository<Card, Long> {
     List<Card> findAllByBoardIdIn(List<Integer> boardIds);
 
     Optional<Card> findByApplicantId(Integer applicantId);
